moved influence model and data
