+++ hrosailing/models/__init__.py
@@ -2,4 +2,9 @@ from hrosailing.models.globe_model import (
     GlobeModel, FlatMercatorProjection, SphericalGlobe
 )
 
+from hrosailing.models.influencemodel import (
+    InfluenceException, InfluenceModel, IdentityInfluenceModel,
+    WindAngleCorrectingInfluenceModel
+)
+
 import hrosailing.models.modelfunctions
